+++ data-raw/create_datasets.R
@@ -30,3 +30,9 @@ grijs_l  <- "#cccccc"
 # c(blauw, blauw_m, blauw_l, oranje, oranje_m, oranje_l, grijs, grijs_m, grijs_l)
 
 usethis::use_data(blauw, blauw_m, blauw_l, oranje, oranje_m, oranje_l, grijs, grijs_m, grijs_l, overwrite = TRUE)
+
+ws_grens_rd <- sf::st_read("data-raw/ws_grens.gpkg", crs = 28992, quiet = TRUE)
+ws_grens_wgs <- sf::st_read("data-raw/ws_grens.gpkg", crs = 28992, quiet = TRUE) |>
+  sf::st_transform(crs = 4326)
+
+usethis::use_data(ws_grens_rd, ws_grens_wgs)